Victor Vasquez at 02/10 - Implementación de metodos controller para listar y registrar productos vía cliente.

+++ BankingProductClient/src/main/java/com/example/bankingproductclient/domain/model/Customer.java
@@ -1,10 +1,8 @@
 package com.example.bankingproductclient.domain.model;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
@@ -27,7 +25,7 @@ import java.util.List;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @Column(name = "email", length = 25, nullable = false)
     @NotNull(message = "The email should not be empty")
     @Email
